Extract out `Data.Percentage` from `Data.Quantity`.
