+++ lib/wallet/integration/src/Test/Integration/Scenario/API/Byron/Network.hs
@@ -14,9 +14,11 @@ import Prelude
 import Cardano.Wallet.Api.Types
     ( ApiNetworkParameters (..)
     )
+import Data.Percentage
+    ( mkPercentage
+    )
 import Data.Quantity
     ( Quantity (..)
-    , mkPercentage
     )
 import Data.Ratio
     ( (%)
