+++ lib/wallet/src/Cardano/Wallet/Gen.hs
@@ -75,11 +75,11 @@ import Crypto.Hash
 import Data.Maybe
     ( fromMaybe
     )
+import Data.Percentage
+    ( Percentage (..)
+    )
 import Data.Proxy
     ( Proxy (..)
-    )
-import Data.Quantity
-    ( Percentage (..)
     )
 import Data.Ratio
     ( denominator
