+++ lib/wallet/test/unit/Cardano/Pool/RankSpec.hs
@@ -32,7 +32,7 @@ import Data.ByteArray.Encoding
     ( Base (Base16)
     , convertToBase
     )
-import Data.Quantity
+import Data.Percentage
     ( clipToPercentage
     )
 import Test.Hspec
